Fix code generator to not use %?
The status codes were being emitted with %?. Not only is this a bad idea
in general, but a change today made it start printing out the integral
type suffixes, so 200 became 200u. Unfortunately, the arg that it's
matching against is a u16, not a uint, so this caused a type conflict.
Using %u is not only safe, but avoids the conflict.

Also add *.dSYM/ to the .gitignore.